US331212: Renamed back classes that had SessionType to be SessionRequest
