+++ checkout/access-checkout-android/src/test/java/com/worldpay/access/checkout/api/session/serialization/CardSessionRequestSerializerTest.kt
@@ -14,19 +14,26 @@ import kotlin.test.assertFailsWith
 
 @RunWith(RobolectricTestRunner::class)
 @Config(manifest = Config.NONE)
-class CVVSessionTypeSerializerTest {
+class CardSessionRequestSerializerTest {
 
-    private val sessionRequestSerializer: Serializer<SessionRequest> = CVVSessionRequestSerializer()
+    private val sessionRequestSerializer: Serializer<SessionRequest> = CardSessionRequestSerializer()
 
     @Test
-    fun `should be able to serialize cvv session request to json`() {
-        val sessionRequest = CVVSessionRequest(
+    fun `should be able to serialize card session request to json`() {
+        val sessionRequest = CardSessionRequest(
+            cardNumber = "0000111122223333",
+            cardExpiryDate = CardExpiryDate(12, 2020),
             cvv = "123",
             identity = "MERCHANT-123"
         )
 
         val expectedRequest = """
                 {
+                    "cardNumber": "0000111122223333",
+                    "cardExpiryDate": {
+                        "month": 12,
+                        "year": 2020
+                    },
                     "cvc": "123",
                     "identity": "MERCHANT-123"
                 }"""
@@ -38,18 +45,8 @@ class CVVSessionTypeSerializerTest {
 
     @Test
     fun `should fail when trying to serialise a cvv session request`() {
-        val sessionRequest = CardSessionRequest(
-            cardNumber = "0000111122223333",
-            cardExpiryDate = CardExpiryDate(12, 2020),
-            cvv = "123",
-            identity = "MERCHANT-123"
-        )
-
-        assertFailsWith<IllegalArgumentException> {
-            sessionRequestSerializer.serialize(
-                sessionRequest
-            )
-        }
+        val sessionRequest = CVVSessionRequest("123", "MERCHANT-123")
+        assertFailsWith<IllegalArgumentException> { sessionRequestSerializer.serialize(sessionRequest) }
     }
 
     private fun removeWhitespace(string: String): String {
