+++ checkout/sample/src/androidTestMock/java/com/worldpay/access/checkout/card/CardSessionRequestMockIntegrationTest.kt
@@ -34,7 +34,7 @@ import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
 @LargeTest
-class CardSessionTypeMockIntegrationTest : AbstractCardFlowUITest() {
+class CardSessionRequestMockIntegrationTest : AbstractCardFlowUITest() {
 
     private val amexCard = "343434343434343"
     private val amexCvv = "1234"
